+++ examples/todo/components/StaticFooter.tsx
@@ -1,0 +1,18 @@
+import * as React from 'react'
+
+export const StaticFooter = () => (
+  <footer className="info">
+    <p>Double-click to edit a todo</p>
+    <p>
+      Template by <a href="http://sindresorhus.com">Sindre Sorhus</a>
+    </p>
+    <p>
+      Created by <a href="https://github.com/merisbahti">Meris</a>
+      {' and '}
+      <a href="http://blog.krawaller.se">David</a>
+    </p>
+    <p>
+      Part of <a href="http://todomvc.com">TodoMVC</a>
+    </p>
+  </footer>
+)
