+++ examples/todo/types.ts
@@ -5,10 +5,10 @@ export const FilterTypeIO = io.union([
   io.literal('all'),
 ])
 export const TodoTypeIO = io.type({ task: io.string, checked: io.boolean })
-export const TodoListTypeIO = io.type({
+export const TodoAppTypeIO = io.type({
   filter: FilterTypeIO,
   todos: io.array(TodoTypeIO),
 })
 export type TodoType = typeof TodoTypeIO._A
 export type FilterType = typeof FilterTypeIO._A
-export type TodoListAtomType = typeof TodoListTypeIO._A
+export type TodoAppType = typeof TodoAppTypeIO._A
